deepface_server/requirements.txt
deepface>=0.0.89
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
